3rd problem set
Due to the computer science olympiad at the department of mathematics I had to do some exercises on the raw text file and unfortunately I couldn't test if they work as I intended. In the pruning exercise I chose value as the pruning treshold not to add more variables to the function.